+++ rust/operator-binary/src/discovery.rs
@@ -75,14 +75,17 @@ pub async fn build_discovery_configmaps(
         .namespace
         .as_deref()
         .context(NoNamespaceSnafu)?;
+    let cluster_domain = &client.kubernetes_cluster_info.cluster_domain;
     let mut discovery_configmaps = vec![build_discovery_configmap(
         name,
         owner,
         hive,
         resolved_product_image,
         chroot,
-        // TODO: make domain configurable
-        vec![(format!("{name}.{namespace}.svc.cluster.local"), HIVE_PORT)],
+        vec![(
+            format!("{name}.{namespace}.svc.{cluster_domain}"),
+            HIVE_PORT,
+        )],
     )?];
 
     // TODO: Temporary solution until listener-operator is finished